requirements.txt
# Backend Improvements - Requirements

# Core Django
Django>=5.0,<6.0
django-environ>=0.11.0

# Django REST Framework
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# CORS Headers
django-cors-headers>=4.3.0

# API Documentation
drf-spectacular>=0.27.0

# HTML Sanitization
bleach>=6.1.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# HTTP Requests
requests>=2.31.0

# Database
psycopg2-binary>=2.9.9  # PostgreSQL adapter

# Optional but recommended
# celery>=5.3.0  # For async tasks
# redis>=5.0.0  # For caching and Celery broker  
# django-redis>=5.4.0  # Django cache backend

# Testing
pytest>=7.4.0
pytest-django>=4.7.0
factory-boy>=3.3.0
